fixing some circular import issues
